updated code

+++ src/app/services/helper.service.ts
@@ -9,7 +9,6 @@ export class HelperService {
   private token = null;
   context = null;
   insertedData = null;
-  paragraphsRange = [];
   constructor(private http: HttpClient) { }
 
   set appToken(token) {
@@ -60,17 +59,24 @@ export class HelperService {
         this.insertedData.load();
         this.insertedData.paragraphs.load();
         this.insertedData.inlinePictures.load();
+        context.document.body.paragraphs.load()
         await context.sync();
         for (const item of this.insertedData.inlinePictures.items) {
           item.delete();
         }
-        for (const item of this.paragraphsRange) {
-          item.delete();
+        for (const item of context.document.body.paragraphs.items) {
+          item.load();
+        }
+        await context.sync();
+        for (const item of context.document.body.paragraphs.items) {
+          const validate = ['First Name: ', 'Last Name: ', 'Email: '];
+          if(validate.find(i => item.text.indexOf(i)>=0)) {
+            item.delete();
+          }
         }
         await context.sync();
         documentLocation = this.insertedData;
       }
-      this.paragraphsRange = [];
       const htmlString = `<img src="${data.avatar}" alt="User Pic">
       <p>First Name: ${data.first_name}</p>
       <p>Last Name: ${data.last_name}</p>
@@ -81,16 +87,8 @@ export class HelperService {
         Word.InsertLocation.after
       );
       this.insertedData.select('End');
-      this.insertedData.paragraphs.load();
+      
       context.trackedObjects.add(this.insertedData);
-      await context.sync();
-      for (const item of this.insertedData.paragraphs.items) {
-        item.load();
-      }
-      await context.sync();
-      for (const item of this.insertedData.paragraphs.items) {
-        this.paragraphsRange.push(item.getRange());
-      }
       await context.sync();
     });
   }
@@ -108,7 +106,7 @@ export class HelperService {
           for (const item of this.insertedData.inlinePictures.items) {
             item.load();
           }
-          for (const item of this.paragraphsRange) {
+          for (const item of context.document.body.paragraphs.items) {
             item.load();
           }
           await context.sync();
